feat : Update Cart, Order, Payment detail Service

+++ src/main/kotlin/com/teamsparta/delivery_system/domain/dto/MenuDto.kt
@@ -12,7 +12,7 @@ data class MenuDto(
     var category: Category,
     var name: String,
     var explanation: String,
-    var price: Long
+    var price: Int
 
 ) {
